# More optimized HID generation - in one commit instead of two.