Timings below for mapping a simple tool of a collection with 10 datasets in a branch with other logging and optimizations - but the idea is pretty clear I think.

Before HID optimization:

```
galaxy.tools.actions INFO 2018-07-18 11:10:20,164 Added output datasets to history (14.193 ms)
...
galaxy.tools.actions INFO 2018-07-18 11:10:20,416 Added output datasets to history (17.583 ms)
...
galaxy.tools.actions INFO 2018-07-18 11:10:20,675 Added output datasets to history (19.875 ms)
...
galaxy.model.orm.engine_factory INFO 2018-07-18 11:10:21,734 Executed [445] SQL requests in for web request [/api/tools] (1036.418676ms)
```

After HID optimization:

```
galaxy.tools.actions INFO 2018-07-18 11:13:41,855 Added output datasets to history (6.049 ms)
...
galaxy.tools.actions INFO 2018-07-18 11:13:42,018 Added output datasets to history (7.192 ms)
...
galaxy.tools.actions INFO 2018-07-18 11:13:42,184 Added output datasets to history (10.277 ms)
...
galaxy.tools.actions INFO 2018-07-18 11:13:42,344 Added output datasets to history (5.800 ms)
...
galaxy.model.orm.engine_factory INFO 2018-07-18 11:13:43,179 Executed [434] SQL requests in for web request [/api/tools] (818.706274ms)
```